// src/web/src/riva/projects/riva-shared/src/lib/layout/enums/layout-sidebar-background-color.enum.ts
export enum LayoutSidebarBackgroundColor {
    Black = 'black',
    Danger = 'danger',
    Flickr = 'flickr',
    IbizaSunset = 'ibiza-sunset',
    Info = 'info',
    KingYna = 'king-yna',
    ManOfSteel = 'man-of-steel',
    Mint = 'mint',
    Primary = 'primary',
    PurpleBliss = 'purple-bliss',
    PurpleLove = 'purple-love',
    Success = 'success',
    White = 'white',
    Warning = 'warning',

    // For transparent layout
    BgGlass1 = 'bg-glass-1',
    BgGlass2 = 'bg-glass-2',
    BgGlass3 = 'bg-glass-3',
    BgGlass4 = 'bg-glass-4',
    BgGlassBlueLagoon = 'bg-glass-blue-lagoon',
    BgGlassElectricViolet = 'bg-glass-electric-violet',
    BgGlassHibiscus = 'bg-glass-hibiscus',
    BgGlassProtage = 'bg-glass-protage',
    BgGlassPurplePizzazz = 'bg-glass-purple-pizzazz',
    BgGlassTundora = 'bg-glass-tundora'
}